move tests and add MAX test
