+++ tests/pytests/test.py
@@ -1378,83 +1378,6 @@ def testNumericRange(env):
             'ft.search', 'idx', 'hello kitty @score:[-inf +inf]', "nocontent")
         env.assertEqual(100, res[0])
 
-def testSuggestions(env):
-    r = env
-    r.expect('ft.SUGADD', 'ac', 'hello world', 1).equal(1)
-    r.expect('ft.SUGADD', 'ac', 'hello world', 1, 'INCR').equal(1)
-
-    res = r.execute_command("FT.SUGGET", "ac", "hello")
-    env.assertEqual(1, len(res))
-    env.assertEqual("hello world", res[0])
-
-    terms = ["hello werld", "hallo world",
-             "yellow world", "wazzup", "herp", "derp"]
-    sz = 2
-    for term in terms:
-        r.expect('ft.SUGADD', 'ac', term, sz - 1).equal(sz)
-        sz += 1
-
-    for _ in r.retry_with_rdb_reload():
-        r.expect('ft.SUGLEN', 'ac').equal(7)
-
-        # search not fuzzy
-        r.expect("ft.SUGGET", "ac", "hello").equal(["hello world", "hello werld"])
-
-        # print  r.execute_command("ft.SUGGET", "ac", "hello", "FUZZY", "MAX", "1", "WITHSCORES")
-        # search fuzzy - shuold yield more results
-        r.expect("ft.SUGGET", "ac", "hello", "FUZZY")\
-         .equal(['hello world', 'hello werld', 'yellow world', 'hallo world'])
-
-        # search fuzzy with limit of 1
-        r.expect("ft.SUGGET", "ac", "hello", "FUZZY", "MAX", "1").equal(['hello world'])
-
-        # scores should return on WITHSCORES
-        res = r.execute_command("ft.SUGGET", "ac", "hello", "WITHSCORES")
-        env.assertEqual(4, len(res))
-        env.assertTrue(float(res[1]) > 0)
-        env.assertTrue(float(res[3]) > 0)
-
-    r.expect("ft.SUGDEL", "ac", "hello world").equal(1L)
-    r.expect("ft.SUGDEL", "ac", "world").equal(0L)
-
-    r.expect("ft.SUGGET", "ac", "hello").equal(['hello werld'])
-
-def testSuggestErrors(env):
-    env.expect('ft.SUGADD ac olah 1').equal(1)
-    env.expect('ft.SUGADD ac olah 1 INCR').equal(1)
-    env.expect('ft.SUGADD ac missing').error().contains("wrong number of arguments")
-    env.expect('ft.SUGADD ac olah not_a_number').error().contains("invalid score")
-    env.expect('ft.SUGADD ac olah 1 PAYLOAD').error().contains('Invalid payload: Expected an argument, but none provided')
-    env.expect('ft.SUGADD ac olah 1 REDIS PAYLOAD payload').error().contains('Unknown argument `REDIS`')
-    env.expect('ft.SUGGET ac olah FUZZ').error().contains("Unrecognized argument: FUZZ")
-    query = 'verylongquery'
-    for _ in range(3):
-        query += query
-    env.expect('ft.SUGGET ac', query).error().contains("Invalid query")
-    env.expect('ft.SUGGET ac', query + query).error().contains("Invalid query length")
-
-def testSuggestPayload(env):
-    r = env
-    env.assertEqual(1, r.execute_command(
-        'ft.SUGADD', 'ac', 'hello world', 1, 'PAYLOAD', 'foo'))
-    env.assertEqual(2, r.execute_command(
-        'ft.SUGADD', 'ac', 'hello werld', 1, 'PAYLOAD', 'bar'))
-    env.assertEqual(3, r.execute_command(
-        'ft.SUGADD', 'ac', 'hello nopayload', 1, 'PAYLOAD', ''))
-    env.assertEqual(4, r.execute_command(
-        'ft.SUGADD', 'ac', 'hello nopayload2', 1))
-
-    res = r.execute_command("FT.SUGGET", "ac", "hello", 'WITHPAYLOADS')
-    env.assertListEqual(['hello world', 'foo', 'hello werld', 'bar', 'hello nopayload', None, 'hello nopayload2', None],
-                         res)
-    res = r.execute_command("FT.SUGGET", "ac", "hello")
-    env.assertListEqual(['hello world',  'hello werld', 'hello nopayload', 'hello nopayload2'],
-                         res)
-    res = r.execute_command(
-        "FT.SUGGET", "ac", "hello", 'WITHPAYLOADS', 'WITHSCORES')
-    # we don't compare the scores beause they may change
-    env.assertEqual(12, len(res))
-
 def testPayload(env):
     r = env
     env.expect('ft.create', 'idx', 'ON', 'HASH', 'PAYLOAD_FIELD', '__payload', 'schema', 'f', 'text').ok()
@@ -2397,12 +2320,6 @@ def testIssue_884(env):
     env.assertEquals(len(expected), len(res))
     for v in expected:
         env.assertContains(v, res)
-
-def testIssue_866(env):
-    env.expect('ft.sugadd', 'sug', 'test123', '1').equal(1)
-    env.expect('ft.sugadd', 'sug', 'test456', '1').equal(2)
-    env.expect('ft.sugdel', 'sug', 'test').equal(0)
-    env.expect('ft.sugget', 'sug', '').equal(['test123', 'test456'])
 
 def testIssue_848(env):
     env.expect('FT.CREATE', 'idx', 'ON', 'HASH', 'SCHEMA', 'test1', 'TEXT', 'SORTABLE').equal('OK')
